refactor: share LLM provider construction and tool schema derivation
