+++ internal/runtime/builtin/chat/executor_test.go
@@ -186,45 +186,6 @@ func TestExecutor_GetMessages(t *testing.T) {
 	})
 }
 
-func TestNormalizeEnvVarExpr(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name     string
-		input    string
-		expected string
-	}{
-		{
-			name:     "EmptyString",
-			input:    "",
-			expected: "",
-		},
-		{
-			name:     "PlainVariableName",
-			input:    "OPENAI_API_KEY",
-			expected: "${OPENAI_API_KEY}",
-		},
-		{
-			name:     "DollarPrefix",
-			input:    "$ANTHROPIC_KEY",
-			expected: "${ANTHROPIC_KEY}",
-		},
-		{
-			name:     "BracedFormat",
-			input:    "${MY_API_KEY}",
-			expected: "${MY_API_KEY}",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			result := normalizeEnvVarExpr(tc.input)
-			assert.Equal(t, tc.expected, result)
-		})
-	}
-}
-
 func TestNewChatExecutor(t *testing.T) {
 	t.Parallel()
